Remove commit presence check when pushing
This can cause unexpected results when trying to push git objects. For example
this will prevent you from pushing a branch that has new commits to a remote
that already has this branch, but pointing to a previous commit. Anyway, git
is smart at pushing objects so there's little value in doing this check.